Moved common color test functions to utils
